// src/model/request.ts
export type customerrequest = {
    name?: string;
    id?: string | number;
    DateOfOrder?: Date;
    Name?: string;
    Email?: string;
    Phone?: string;
    Address?: string;
    quantity?: number;
    orderdetails?: string;
    price?: number;
    city?: string;
    state?: String;
    zipcode?: number;
}